delete decorators.pyi
