+++ stubs/cachetools/cachetools/__init__.pyi
@@ -6,6 +6,18 @@ from .ttl import TTLCache as TTLCache
 
 from typing import Callable, Generic, Iterator, TypeVar
 
+from _typeshed import IdentityFunction
+from typing import Any, Callable, ContextManager, MutableMapping, TypeVar
+
+_KT = TypeVar("_KT")
+
+def cached(
+    cache: MutableMapping[_KT, Any] | None, key: Callable[..., _KT] = ..., lock: ContextManager[Any] | None = ...
+) -> IdentityFunction: ...
+def cachedmethod(
+    cache: Callable[[Any], MutableMapping[_KT, Any] | None], key: Callable[..., _KT] = ..., lock: ContextManager[Any] | None = ...
+) -> IdentityFunction: ...
+
 _KT = TypeVar("_KT")
 _VT = TypeVar("_VT")
